Fixed logical error in computing Mma k slice.
Had a strange logical error where I was loading the wrong k slices for
the Mma into registers for the computation. Was doing different things
wrong for A and B. Had a > instead of a >= or a < and a logical flip of
the outputs. MMA seems to be mostly working, although still some issues
at the end of the computation.