Changed global array to half2*. Wrote squared sum.
Decided it made more sense to treat the global array as a half2 array
since this is fundamentally what it is. It is a little more complicated
than that since it is technically half, but since it is a better idea to
read half2 at a time I decided to leave it as half2. Wrote a basic
method to compute the sum of the squares of each dimensions of each
point.