Unrolling loops forced inspect back into registers
When I added a conditional check to see if we were inside of the actual
points we were inspecting, my arrays were moved into local memory. This
is because the compiler wasn't have an easy job determining array
indexing. Forcing the unroll of all three loops forced them all to be
computed, eliminating the need to compute indexes. Performance has been
restored.