Created PointGenerator abstraction.
Factored out the generation of points from the rest of the code. Pulled
current file reading code into its own FilePointGenerator class. Next
step will be to create an exponential distribution generator class as
well. The point padding is all done in the PointListBuilder, as that is
separate from generating the real points.