Overhauled parameters passed into kernel and host.
Separated parameters required to run from the host's perspective from the device's perspective. This made it easier to report all of the results back from the findPairs() routine. Lots of code moved out of main.cu and into findPairs(), as it handles all the required device allocations, and returns the complete results. Timing and measurement is now all done inside of findPairs().