Sum squared's producing incorrect results.
Tensor cores apparently round towards zero when they accumulate. I attempted to recreate this by using a single thread to accumulate each point. This way the ordering is preserved, and we use round to zero mode in this accumulation. AtomicAdd() does not allow for round to zero, only nearest. I could potentially do a real reduction in shared memory using round to zero mode though to speed this up. I don't think it is the slowest part though.