Removed atomicAdd code in sumSquared kernel.
Can't use atomicAdd because it rounds to nearest even, which is different than tensor cores. I get worse results it seems if I use it. When I count pairs, I should always get the same number or pairs as points for the self join. If I round up, I notice that for higher dimensional datasets, the distance between self pairs gradually increases as we accumulate more and more error due to rounding. The tensor cores seem to underestimate distance, while this kernel overestimates it.