Added reRun method to cuda code.
Can rerun a search, using the previous inputs in global memory. Added
the ability to do this from python. It isn't very clean as of now,
however it seems to be working well. Need to be careful to clean up
resources after use with this scheme. Experiments now run significantly
faster.

Also refactored variable names in python to be PEP8 compliant.