Created a rerunnable pairs finder algorithm.
It is just a class that is callable, and the first time it is run, it
invokes the first method (which reads/generates the dataset), and on
subsequent times it inokes the second method (This just reruns with the
memory on the GPU).

Changed the code to take in this simpler method. This simplified the
logic greatly, as users of the algorithm do not need to worry about if
the dataset has already been downloaded.