Gathered sift and tiny tedjoin results.
Stored the output files for these runs. Added them to the python plots
for comparison. They are significantly slower than my routines. The tiny
implementation required me using all the shared memory on the GPU, which
might have killed its caching performance. It might not be relevant to
actually report this.