Updated accuracy computation for all selectiviies.
Made script more flexible so it can run on raw GDS-Join data. It filters
out header rows until it gets to the actual neighbors.

I added the 128, and 256 selectivity neighbor tables for comparison.

Added the ability to detect if we are running locally, or on slurm. This
makes the script easy to debug, and deploy without any changes.