Fixed html render issues.
Enumerate was bolding the entire text in html. Had to just write soem
text in \textbf{} to make this work.

I accidentally had a forced newline in my paper, which came through in
the html version in a bad spot, making it look very bad. Had to add
optional hyphen break for the paper.